// src/app/interfaces/box.ts
export interface Box {
    batteryStatus: number
    connectivityStatus: number
    detectionCount: number
    id: string
    lastHeartbeat: number
    locationLat: number
    locationLong: number
    name: string
    ping: number
    registeredAt: number
    serialNumber: string
    type: string
}